+++ src/promptflow/promptflow/_sdk/operations/_run_submitter.py
@@ -45,7 +45,7 @@ from promptflow._sdk.entities._run import Run
 from promptflow._sdk.operations._local_storage_operations import LocalStorageOperations
 from promptflow._sdk.operations._run_operations import RunOperations
 from promptflow._utils.context_utils import _change_working_dir
-from promptflow.batch_engine import BatchEngine
+from promptflow.batch import BatchEngine
 from promptflow.contracts.flow import Flow as ExecutableFlow
 from promptflow.contracts.run_info import Status
 from promptflow.contracts.run_mode import RunMode
